feat: implement core FS operations and payload algorithms
